+++ package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
-    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -49,7 +48,8 @@
     "@types/react": "^16.9.44",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.5",
-    "sass": "^1.26.10",
+    "dotenv": "^8.2.0",
+    "node-sass": "^4.14.1",
     "typescript": "^3.9.7"
   }
 }
