+++ public/index.html
@@ -1,24 +1,22 @@
 <!DOCTYPE html>
 <html lang="es">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/icons/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#6705c6" />
-    <meta
-      name="description"
-      content="La mejor plataforma de educacion online"
-    />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/icons/logo192.png" />
-    <link rel="manifest" href="%PUBLIC_URL%/icons/manifest.json" />
 
-    <title>Escuela Virtual</title>
-  </head>
-  <body>
-    <noscript
-      >Porfavor habilita la ejecucion de javascript para que este sitio puea
-      funcionar correctamente</noscript
-    >
-    <div id="root"></div>
-  </body>
+<head>
+	<meta charset="utf-8" />
+	<link rel="icon" href="%PUBLIC_URL%/img/favicon.ico" />
+	<meta name="viewport" content="width=device-width, initial-scale=1" />
+	<meta name="theme-color" content="#6705c6" />
+	<meta name="description" content="La mejor plataforma de educacion online" />
+	<link rel="apple-touch-icon" href="%PUBLIC_URL%/img/logo192.png" />
+	<link rel="manifest" href="%PUBLIC_URL%/img/manifest.json" />
+
+	<title>Escuela Virtual</title>
+</head>
+
+<body>
+	<noscript>Por favor habilita la ejecucion de javascript para que este sitio pueda
+		funcionar correctamente</noscript>
+	<div id="root"></div>
+</body>
+
 </html>
