+++ public/manifest.json
@@ -3,17 +3,17 @@
   "name": "Escuela Virtual",
   "icons": [
     {
-      "src": "./icons/favicon.ico",
+      "src": "./img/favicon.ico",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
     },
     {
-      "src": "./icons/logo192.png",
+      "src": "./img/logo192.png",
       "type": "image/png",
       "sizes": "192x192"
     },
     {
-      "src": "./icons/logo512.png",
+      "src": "./img/logo512.png",
       "type": "image/png",
       "sizes": "512x512"
     }
