+++ public/scripts/pwa.js
