--- src/components/App/style.scss
@@ -1,1 +1,0 @@
-
