+++ src/design/organisms/Navigation/NavAppInfo.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import "./style.scss";
 
 const NavAppInfo = (props: any) => {
   return (
