+++ src/pages/home/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import Navigation from "../../design/organisms/Navigation/index";
+
+export default function Home(): JSX.Element {
+  return <Navigation />;
+}
