+++ src/pages/login/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Login(): JSX.Element {
+  return <h1>Este es el login</h1>;
+}
+
+export default Login;
