--- src/routes/home/index.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import Navigation from "../../components/navigation/index";
-
-const Home = () => {
-  return (
-    <Navigation />
-  )
-}
-
-export default Home;
