--- src/routes/login/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Login = () => {
-  return (
-    <h1>Este es el login</h1>
-  )
-}
-
-export default Login;
