+++ src/styles/base/_base.scss
@@ -1,0 +1,1 @@
+@import "./_normalize.scss";
