+++ src/styles/base/_normalize.scss
@@ -1,5 +1,3 @@
-@import "./variables";
-
 html,
 body {
   padding: 0;
