+++ src/styles/index.scss
@@ -1,0 +1,6 @@
+@import "./utils/_utils.scss";
+@import "./base/_base.scss";
+@import "./pages/_pages.scss";
+@import "./atoms/_atoms.scss";
+@import "./organisms/_organisms.scss";
+@import "./template/_template.scss";
