+++ src/styles/organisms/_nav.scss
@@ -1,9 +1,7 @@
-@import "../../variables";
-
 #Navigation {
   display: flex;
   align-content: center;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
   width: 100%;
   min-height: 20px;
