+++ src/styles/organisms/_organisms.scss
@@ -1,0 +1,3 @@
+@import "./_nav.scss";
+@import "./_page404.scss";
+@import "./_searcher.scss";
