+++ src/styles/organisms/_page404.scss
@@ -1,5 +1,3 @@
-@import "../../variables";
-
 .pageNotFoundContent {
   justify-content: center;
   align-items: center;
