+++ src/styles/organisms/_seeker.scss
