+++ src/styles/utils/_variables.scss
@@ -228,7 +228,7 @@ $smallButtonLight: 12px;
 
 @font-face {
   font-family: "Cascadia Code PL";
-  src: url(./assets/fonts/CascadiaCodePL.ttf);
+  src: url(../../assets/fonts/CascadiaCodePL.ttf);
 }
 
 // typography titles and body
